Fixed an issue with new instances of denormed m2m fields

+++ denorm/denorms.py
@@ -14,28 +14,41 @@ def many_to_many_pre_save(sender, instance, **kwargs):
     """
     Updates denormalised many-to-many fields for the model
     """
-    for m2m in sender._meta.local_many_to_many:
-        # This gets us all m2m fields, so limit it to just those that are denormed
-        if hasattr(m2m, "denorm"):
-            # Does some extra jiggery-pokery for "through" m2m models.
-            # May not work under lots of conditions.
-            if hasattr(m2m.rel, 'through_model'):
+    if instance.pk:
+        # need a primary key to do m2m stuff
+        for m2m in sender._meta.local_many_to_many:
+            # This gets us all m2m fields, so limit it to just those that are denormed
+            if hasattr(m2m, "denorm"):
+                # Does some extra jiggery-pokery for "through" m2m models.
+                # May not work under lots of conditions.
+                if hasattr(m2m.rel, 'through_model'):
             
-                # Clear exisiting through records (bit heavy handed?)
-                kwargs = { m2m.related.var_name: instance, }
+                    # Clear exisiting through records (bit heavy handed?)
+                    kwargs = { m2m.related.var_name: instance, }
             
-                # Can't use m2m_column_name in a filter
-                # kwargs = { m2m.m2m_column_name(): instance.pk, }
-                m2m.rel.through_model.objects.filter(**kwargs).delete()
+                    # Can't use m2m_column_name in a filter
+                    # kwargs = { m2m.m2m_column_name(): instance.pk, }
+                    m2m.rel.through_model.objects.filter(**kwargs).delete()
             
-                values = m2m.denorm.func(instance)
-                for value in values:
-                    kwargs.update({m2m.m2m_reverse_name(): value.pk,})
-                    m2m.rel.through_model.objects.create(**kwargs)
+                    values = m2m.denorm.func(instance)
+                    for value in values:
+                        kwargs.update({m2m.m2m_reverse_name(): value.pk,})
+                        m2m.rel.through_model.objects.create(**kwargs)
 
-            else:
-                values = m2m.denorm.func(instance)
-                setattr(instance, m2m.attname, values)
+                else:
+                    values = m2m.denorm.func(instance)
+                    setattr(instance, m2m.attname, values)
+                
+def many_to_many_post_save(sender, instance, created, **kwargs):
+    if created:
+        def check_resave():
+            for m2m in sender._meta.local_many_to_many:
+                if hasattr(m2m, "denorm"):
+                    return True
+            return False
+                
+        if check_resave():
+            instance.save()
 
 class Denorm(object):
